+++ Tik-Tok Lite/ImagePicker.swift
@@ -5,44 +5,6 @@
 //  Created by Fantom on 19.12.2021.
 //
 
-//import SwiftUI
-//
-//struct ImagePicker: UIViewControllerRepresentable{
-//
-//    class Coordinator: NSObject, UINavigationControllerDelegate, UIImagePickerControllerDelegate{
-//        let parent: ImagePicker
-//
-//        init(_ parent: ImagePicker ) {
-//            self.parent = parent
-//        }
-//        
-//        func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
-//            if let uiImage = info[.originalImage] as? UIImage {
-//                parent.image = uiImage
-//            }
-//            parent.presentationMode.wrappedValue.dismiss()
-//        }
-//    }
-//
-//    @Environment(\.presentationMode) var presentationMode
-//    @Binding var image: UIImage?
-//
-//
-//    func makeCoordinator() -> Coordinator {
-//        Coordinator(self)
-//    }
-//
-//    func makeUIViewController(context: UIViewControllerRepresentableContext<ImagePicker>) ->  UIImagePickerController {
-//        let picker = UIImagePickerController()
-//        picker.delegate = context.coordinator
-//        return picker
-//    }
-//
-//    func updateUIViewController(_ uiViewController: UIImagePickerController, context: UIViewControllerRepresentableContext<ImagePicker>) {
-//
-//    }
-//}
-
 
 import SwiftUI
 import UIKit
